Split IntegrationTest to controller + integration
No new tests were implemented, only existing ones were reused were applicable.
Currently, this means that the newly created FlightDataHandlerImplTest which consists of mix of unit testing of FlightDataHandler but also some integration testing with GUI cannot be used and all of the tests are silenced. This is because GUI components are no longer exposed, and they will not be.